# packages/openapi/skill-registry.openapi.yaml
openapi: 3.1.0
info:
  title: ManagedSkillHub API
  description: Self-hosted skill registry for AI agents.
  version: 0.1.0
servers:
- url: http://localhost:3040
  description: Local API server
...
paths:
...
  /discover:
    get:
      operationId: discover
      tags:
      - Discovery
      summary: Registry discovery endpoint
      description: Returns registry identity, canonical API base URL, entrypoints,
        and runtime auth requirements. May require discoveryBearer when DISCOVERY_AUTH_MODE=bearer
        or an Authentik access token with the advertised discovery scope when DISCOVERY_AUTH_MODE=oidc.
      security:
      - discoveryBearer: []
      - agentSession: []
      - agentOidc:
        - managedskillhub:discovery
      responses:
        '200':
          description: Registry metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /admin/proposals/{proposalId}/judge:
    post:
      operationId: judgeProposal
      tags:
      - Judgements
      summary: Request judgement for a proposal
      security:
      - adminSession: []
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Judgement result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgementResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/judge/file:
    post:
      operationId: judgeFile
      tags:
      - Judgements
      summary: Judge a single file
      security:
      - adminSession: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Judgement result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgementResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/judge/skill/{skillId}/version/{version}:
    post:
      operationId: judgeSkillVersion
      tags:
      - Judgements
      summary: Judge a published skill version
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Judgement result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgementResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/judgements/{targetType}/{targetId}:
    get:
      operationId: listJudgements
      tags:
      - Judgements
      summary: List judgements for a target
      security:
      - adminSession: []
      parameters:
      - name: targetType
        in: path
        required: true
        schema:
          type: string
          enum:
          - proposal
          - skill
          - file
      - name: targetId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: List of judgements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgementListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /admin/proposals/{proposalId}/files/{fileId}/judge:
    post:
      operationId: adminJudgeProposalFile
      tags:
      - Judgements
      summary: Re-run judgement for one stored proposal file
      security:
      - adminSession: []
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Judgement result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgementResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: skill_hub_session
    publicReadBearer:
      type: http
      scheme: bearer
      description: Runtime-configurable bearer token for published skill consumption
        when PUBLIC_READ_AUTH_MODE=bearer.
    proposalBearer:
      type: http
      scheme: bearer
      description: Runtime-configurable bearer token for proposal duplicate check,
        submit, upload, finalization, notices, and status when PROPOSAL_AUTH_MODE=bearer.
    discoveryBearer:
      type: http
      scheme: bearer
      description: Runtime-configurable bearer token for discovery/contract endpoints
        when DISCOVERY_AUTH_MODE=bearer.
    agentOidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.example.invalid/application/o/managedskillhub-agent-device/.well-known/openid-configuration
      description: Runtime Authentik access token. The concrete issuer, client ID,
        scopes, Device Authorization endpoint, token endpoint, and OpenID configuration
        URL must be read from /discover; this static URL is intentionally only a contract
        placeholder.
      x-device-authorization:
        flow: device_code
        runtimeMetadataPath: /discover
        verificationLinkField: verification_uri_complete
        deviceCodeIsSecret: true
    agentSession:
      type: apiKey
      in: header
      name: Authorization
      description: 'Short-lived agent session code sent as "Authorization: AgentSession
        <code>". Created through POST /agent-sessions after proving possession of
        the relevant area bearer tokens.'
  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    DiscoveryResponse:
      type: object
      required:
      - name
      - version
      - description
      - registryId
      - registryName
      - apiBaseUrl
      - readAuthRequired
      - proposalAuthRequired
      - discoveryAuthRequired
      - documentation
      - capabilities
      - entrypoints
      properties:
        name:
          type: string
        registryId:
          type: string
        registryName:
          type: string
        apiBaseUrl:
          type: string
        version:
          type: string
        description:
          type: string
        proposalAuthRequired:
          type: boolean
        discoveryAuthRequired:
          type: boolean
        authSchemes:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeAuthScheme'
        credentialSetupScriptUrl:
          type:
          - string
          - 'null'
        workflowNotes:
          type: object
          properties:
            conversationLanguage:
              type: string
            readPath:
              type: string
            proposalPath:
              type: string
            publishedSkillDownload:
              type: string
            howToProposeUrl:
              type: string
            frontendUrl:
              type: string
            adminOnlyActions:
              type: array
              items:
                type: string
        readAuthRequired:
          type: boolean
        documentation:
          type: object
          required:
          - human
          - openapi
          - frontend
          properties:
            human:
              type: string
              format: uri
            openapi:
              type: string
            frontend:
              type: string
        capabilities:
          type: array
          items:
            type: object
            required:
            - id
            - name
            - description
            properties:
              id:
                type: string
              name:
                type: string
              description:
                type: string
        entrypoints:
          type: array
          items:
            type: object
            required:
            - id
            - name
            - description
            - methods
            - path
            - url
            properties:
              id:
                type: string
              name:
                type: string
              description:
                type: string
              methods:
                type: array
                items:
                  type: string
                  enum:
                  - GET
                  - POST
                  - PATCH
                  - PUT
                  - DELETE
              path:
                type: string
              url:
                type: string
    RuntimeAuthScheme:
      type: object
      required:
      - id
      - type
      - appliesTo
      properties:
        id:
          type: string
        type:
          type: string
          enum:
          - bearer
          - oauth2
          - agent-session
        flow:
          type: string
          enum:
          - device_code
        issuer:
          type: string
          format: uri
        openIdConfigurationUrl:
          type: string
          format: uri
        authorizationEndpoint:
          type: string
          format: uri
        deviceAuthorizationEndpoint:
          type: string
          format: uri
        tokenEndpoint:
          type: string
          format: uri
        clientId:
          type: string
        scopes:
          type: array
          items:
            type: string
        appliesTo:
          type: array
          items:
            type: string
            enum:
            - discovery
            - public-read
            - proposal
...
    JudgementResponse:
      type: object
      required:
      - id
      - targetType
      - targetId
      - dimensions
      - overallRisk
      - summary
      - skillPurposeSummary
      - createdAt
      properties:
        id:
          type: string
        targetType:
          type: string
          enum:
          - proposal
          - skill
          - file
        targetId:
          type: string
        dimensions:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/JudgementDimension'
        overallRisk:
          type: string
          enum:
          - low
          - medium
          - high
          - critical
          - no_judge_available
        summary:
          type: string
        skillPurposeSummary:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    JudgementDimension:
      type: object
      required:
      - risk
      - score
      properties:
        risk:
          type: string
          enum:
          - low
          - medium
          - high
          - critical
        score:
          type: number
        reason:
          type: string
    JudgementListResponse:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/JudgementResponse'
...
    ErrorResponse:
      type: object
      required:
      - error
      - code
      - requestId
      properties:
        error:
          type: string
        code:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
          properties:
            authRequired:
              type: boolean
            authArea:
              type: string
              enum:
              - discovery
              - public-read
              - proposal
            authScheme:
              type: string
              enum:
              - bearer
              - agent-session
            discoverUrl:
              type: string
            credentialSetupScriptUrl:
              type:
              - string
              - 'null'
            recommendation:
              type: string
        originalError:
          type: string
...